third_party/WebKit/LayoutTests/external/wpt/html/semantics/document-metadata/the-base-element/base_srcdoc-expected.txt

This is a testharness.js-based test.
FAIL base element in srcdoc document should resolve against its fallback base URI assert_equals: expected "http://web-platform.test:8001/html/semantics/document-metadata/the-base-element/test" but got ""
Harness: the test ran to completion.
